requirements.txt
PyYAML>=6.0
requests>=2.28.0
tqdm>=4.64.0
xarray>=2023.1.0
numpy>=1.21.0
scikit-learn>=1.3.0
netCDF4>=1.6.0
cdsapi>=0.7.4
